+++ Godeps/_workspace/src/github.com/openshift/source-to-image/pkg/errors/errors.go
@@ -69,7 +69,7 @@ func NewPullImageError(name string, err error) error {
 		Message:    fmt.Sprintf("unable to get %s", name),
 		Details:    err,
 		ErrorCode:  PullImageError,
-		Suggestion: "check image name, or if using local image add --force-pull=false flag",
+		Suggestion: "check image name, or if using local image add --pull-policy=never flag",
 	}
 }
