--- Godeps/_workspace/src/github.com/openshift/source-to-image/pkg/util/child.go
@@ -1,30 +1,0 @@
-// +build !linux
-
-/*
-Derived from:
-
-https://github.com/kubernetes/kubernetes
-
-Copyright 2016 The Kubernetes Authors All rights reserved.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package util
-
-import (
-	"os"
-)
-
-// childSignals are the allowed signals that can be sent to children in Windows to terminate
-var childSignals = []os.Signal{os.Interrupt}
